refactor: 이미지 업로드 로직을 유틸리티 함수로 분리
